完成创建考试用例

+++ manage-parent/exam-manage-biz/src/main/java/com/kalix/exam/manage/biz/ExamExamineeBeanServiceImpl.java
@@ -1,25 +1,37 @@
 package com.kalix.exam.manage.biz;
 
+import com.kalix.admin.core.api.biz.IOrganizationBeanService;
 import com.kalix.admin.core.api.biz.IUserBeanService;
+import com.kalix.admin.core.dto.model.OrganizationDTO;
 import com.kalix.admin.core.entities.UserBean;
+import com.kalix.exam.manage.api.biz.IExamCreateBeanService;
 import com.kalix.exam.manage.api.biz.IExamExamineeBeanService;
 import com.kalix.exam.manage.api.dao.IExamExamineeBeanDao;
 import com.kalix.exam.manage.dto.ExamOrgDto;
+import com.kalix.exam.manage.entities.ExamCreateBean;
 import com.kalix.exam.manage.entities.ExamExamineeBean;
 import com.kalix.framework.core.api.persistence.JsonStatus;
-import com.kalix.framework.core.impl.biz.GenericBizServiceImpl;
+import com.kalix.framework.core.impl.biz.ShiroGenericBizServiceImpl;
 
 import javax.persistence.Transient;
-import java.util.Collection;
-import java.util.List;
-import java.util.stream.Collectors;
+import java.util.*;
 
-public class ExamExamineeBeanServiceImpl extends GenericBizServiceImpl<IExamExamineeBeanDao,ExamExamineeBean> implements IExamExamineeBeanService {
+public class ExamExamineeBeanServiceImpl extends ShiroGenericBizServiceImpl<IExamExamineeBeanDao,ExamExamineeBean> implements IExamExamineeBeanService {
 
     private IUserBeanService userBeanService;
+    private IExamCreateBeanService examCreateBeanService;
+    private IOrganizationBeanService organizationBeanService;
 
     public void setUserBeanService(IUserBeanService userBeanService) {
         this.userBeanService = userBeanService;
+    }
+
+    public void setExamCreateBeanService(IExamCreateBeanService examCreateBeanService) {
+        this.examCreateBeanService = examCreateBeanService;
+    }
+
+    public void setOrganizationBeanService(IOrganizationBeanService organizationBeanService) {
+        this.organizationBeanService = organizationBeanService;
     }
 
     @Override
@@ -36,34 +48,90 @@ public class ExamExamineeBeanServiceImpl extends GenericBizServiceImpl<IExamExam
         }
 
         try {
-            // 删除examId相关信息
-            deleteByExamId(examId);
-
             // 批量插入考生信息
+            List<ExamExamineeBean> examExamineans = new ArrayList<>();
+            List<Long> userIds = new ArrayList<>();
+
             if (orgIds.indexOf(",") != -1) {
                 String[] orgIdArr = orgIds.split(",");
-
                 if (orgIdArr.length > 0) {
                     for (String orgId : orgIdArr) {
                         if (orgId != null) {
-                            List<UserBean> users = userBeanService.findUserByOrgId(Long.parseLong(orgId.trim()));
-                            List<Long> userIds = users.stream().map((u)->u.getId()).collect(Collectors.toList());
-
+                            setExamineeUserId(examId, orgId, userIds, examExamineans);
                         }
                     }
                 }
+            } else {
+                setExamineeUserId(examId, orgIds, userIds, examExamineans);
             }
-            jsonStatus.setMsg("考生分配成功");
-            jsonStatus.setSuccess(true);
+            if (examExamineans.size() > 0) {
+                // 删除examId相关信息
+                deleteByExamId(examId);
+                // 批量添加分配的考生
+                dao.addBatch(examExamineans);
+                // 修改考试分配状态
+                updateDistributeStat(examId);
+                // 发通知消息
+
+                jsonStatus.setMsg("考生分配成功");
+                jsonStatus.setSuccess(true);
+            } else {
+                jsonStatus.setMsg("请先在考生所在单位添加考生");
+                jsonStatus.setFailure(true);
+            }
         } catch(Exception e) {
             jsonStatus.setMsg("考生分配失败");
-            jsonStatus.setSuccess(true);
+            jsonStatus.setFailure(true);
         }
         return jsonStatus;
+    }
+
+    private void setExamineeUserId(Long examId, String orgId, List<Long> userIds, List<ExamExamineeBean> examExamineans) {
+        List<UserBean> users = userBeanService.findUserByOrgId(Long.parseLong(orgId.trim()));
+        Long currentUserId = shiroService.getCurrentUserId();
+        if (users != null) {
+            ExamExamineeBean examExamineeBean = null;
+            for (UserBean user : users) {
+                if (!userIds.contains(user.getId())) {
+                    userIds.add(user.getId());
+                    examExamineeBean = new ExamExamineeBean();
+                    examExamineeBean.setExamId(examId);
+                    examExamineeBean.setState("未考");
+                    examExamineeBean.setUserId(user.getId());
+                    examExamineeBean.setOrgId(Long.parseLong(orgId));
+                    examExamineeBean.setCreateById(currentUserId);
+                    examExamineeBean.setUpdateById(currentUserId);
+                    examExamineeBean.setCreationDate(new Date());
+                    examExamineans.add(examExamineeBean);
+                }
+            }
+        }
     }
 
     @Override
     public void deleteByExamId(Long examId) {
         dao.updateNativeQuery("delete from exam_examinee where examid=" + examId);
     }
+
+    @Override
+    public Map<String, Object> getExamOrgTree(Long examId) {
+        OrganizationDTO organizationDTO = organizationBeanService.getAllOrg(null);
+        List<Long> orgIds = getOrgIdsByExamId(examId);
+        Map<String, Object> map = new HashMap<>();
+        map.put("treeData", organizationDTO);
+        map.put("orgIds", orgIds);
+        return map;
+    }
+
+    private void updateDistributeStat(Long examId) {
+        ExamCreateBean examCreateBean = examCreateBeanService.getEntity(examId);
+        examCreateBean.setDistributeStat("已分配");
+        examCreateBeanService.updateEntity(examCreateBean);
+    }
+
+    private List<Long> getOrgIdsByExamId(Long examId) {
+        return dao.findByNativeSql("select orgid from exam_examinee where examid=?", Long.class, examId);
+    }
+
+
 }
